Generic AuthenticationNotification component which is used when showing event/registration warnings
