+++ src/domain/app/authenticationNotification/authenticationNotification.module.scss
@@ -1,4 +1,4 @@
-.authRequiredNotification {
+.authenticationNotification {
   margin: 0 0 var(--spacing-m) 0;
   border-top-width: var(--notification-border-width);
   border-left-width: 0;
