+++ src/domain/app/i18n/fi.json
@@ -19,7 +19,7 @@
     "some": "Facebook, Google tai Twitter tunnus",
     "yle": "YLE-tunnus"
   },
-  "authRequiredNotification": { 
+  "authenticationNotification": { 
     "button": "kirjautua sisään",
     "part1": "Voit jatkaa palveluun tutustumista, mutta sinun täytyy",
     "part2": ", jos haluat tallentaa muutoksia."
